refactor: move review report function from model to utils
